+++ clinvar/tests/factories.py
@@ -1,21 +1,8 @@
 """Factory Boy factory classes for ``clinvar``."""
-
+import attr
 import factory
 
-from variants.tests.factories_data import SMALL_VARS, small_var_iterator, small_var_attribute
-
 from ..models import Clinvar
-
-#: The gene to create the frequencies for.
-SYMBOL = "LAMA1"
-
-#: Base data for ClinVar by gene.
-CLINVAR_VARS = {SYMBOL: {"strand": "+"}}
-
-
-def clin_var_attribute(gene, attribute):
-    """Return attribute value for the given gene and attribute"""
-    return CLINVAR_VARS[gene][attribute]
 
 
 class ClinvarFactory(factory.django.DjangoModelFactory):
@@ -24,31 +11,26 @@ class ClinvarFactory(factory.django.DjangoModelFactory):
     class Meta:
         model = Clinvar
 
-    release = small_var_attribute(SYMBOL, "release")
-    chromosome = small_var_attribute(SYMBOL, "chromosome")
-    position = small_var_iterator(SYMBOL, "position")
-    reference = small_var_iterator(SYMBOL, "reference")
-    alternative = small_var_iterator(SYMBOL, "alternative")
+    release = "GRCh37"
+    chromosome = factory.Iterator((list(map(str, range(1, 23))) + ["X", "Y"]))
+    position = factory.Sequence(lambda n: n * 100)
+    reference = factory.Iterator("ACGT")
+    alternative = factory.Iterator("CGTA")
+    start = factory.SelfAttribute("position")
+    stop = factory.LazyAttribute(lambda o: o.position + len(o.reference) - len(o.alternative))
 
-    start = small_var_iterator(SYMBOL, "position")
-    stop = factory.Sequence(
-        lambda n: SMALL_VARS[SYMBOL]["position"][n]
-        + len(SMALL_VARS[SYMBOL]["reference"][n])
-        - len(SMALL_VARS[SYMBOL]["alternative"][n])
-    )
-
-    strand = clin_var_attribute(SYMBOL, "strand")
+    strand = "+"
     variation_type = "Variant"
     variation_id = factory.Sequence(lambda n: 12345 + n)
     rcv = factory.Sequence(lambda n: "RCV%d" % (12345 + n))
     scv = factory.Sequence(lambda n: ["RCV%d" % (12345 + n)])
     allele_id = factory.Sequence(lambda n: 12345 + n)
-    symbol = SYMBOL
-    hgvs_c = small_var_iterator(SYMBOL, "refseq_hgvs_c")
-    hgvs_p = small_var_iterator(SYMBOL, "refseq_hgvs_p")
+    symbol = factory.Sequence(lambda n: "SYMBOL%d" % n)
+    hgvs_c = "c.123C>T"
+    hgvs_p = "p.I2T"
     molecular_consequence = "some-molecular-consequence"
-    clinical_significance = "benign"
-    clinical_significance_ordered = ["benign"]
+    clinical_significance = ""
+    clinical_significance_ordered = []
     pathogenic = 0
     likely_pathogenic = 0
     uncertain_significance = 0
@@ -69,3 +51,59 @@ class ClinvarFactory(factory.django.DjangoModelFactory):
     xrefs = ["Some xref"]
     dates_ordered = ["2016-06-14"]
     multi = 1
+
+
+@attr.s(auto_attribs=True)
+class ClinvarFormDataFactory:
+    # Testbase should add %s_gt % patient !!!
+    # no need for this data outside clinvar queries.
+    # Pathogenicity
+    clinvar_include_benign: bool = False
+    clinvar_include_likely_benign: bool = False
+    clinvar_include_uncertain_significance: bool = False
+    clinvar_include_likely_pathogenic: bool = False
+    clinvar_include_pathogenic: bool = False
+    # Origin
+    clinvar_origin_germline: bool = True
+    clinvar_origin_somatic: bool = False
+    # Reliability
+    clinvar_status_practice_guideline: bool = True
+    clinvar_status_expert_panel: bool = True
+    clinvar_status_multiple_no_conflict: bool = True
+    clinvar_status_conflict: bool = True
+    clinvar_status_single: bool = True
+    clinvar_status_no_criteria: bool = True
+    clinvar_status_no_assertion: bool = True
+    # Database
+    database_select: str = "refseq"
+    # Genotypes
+    compound_recessive_enabled: bool = False
+    # Limit on number of rows
+    result_rows_limit: int = 500
+    # Filter selection for form.
+    flag_bookmarked: bool = True
+    flag_candidate: bool = True
+    flag_final_causative: bool = True
+    flag_for_validation: bool = True
+    flag_phenotype_match_empty: bool = True
+    flag_phenotype_match_negative: bool = True
+    flag_phenotype_match_positive: bool = True
+    flag_phenotype_match_uncertain: bool = True
+    flag_simple_empty: bool = True
+    flag_summary_empty: bool = True
+    flag_summary_negative: bool = True
+    flag_summary_positive: bool = True
+    flag_summary_uncertain: bool = True
+    flag_validation_empty: bool = True
+    flag_validation_negative: bool = True
+    flag_validation_positive: bool = True
+    flag_validation_uncertain: bool = True
+    flag_visual_empty: bool = True
+    flag_visual_negative: bool = True
+    flag_visual_positive: bool = True
+    flag_visual_uncertain: bool = True
+    # Clinvar/HGMD
+    require_in_clinvar: bool = False
+    remove_if_in_dbsnp: bool = False
+    require_in_hgmd_public: bool = False
+    display_hgmd_public_membership: bool = True