+++ geneinfo/tests/factories.py
@@ -2,11 +2,28 @@
 
 import factory
 
-from ..models import Hgnc, Hpo, Mim2geneMedgen
+from ..models import Hgnc, Hpo, Mim2geneMedgen, Acmg, RefseqToHgnc
+
+
+class RefseqToHgncFactory(factory.django.DjangoModelFactory):
+    """Factory for the ``RefseqToHgnc`` model."""
+
+    class Meta:
+        model = RefseqToHgnc
+
+    entrez_id = factory.Sequence(lambda n: str(n))
+    hgnc_id = factory.Sequence(lambda n: "HGNC:%d" % n)
 
 
 class HgncFactory(factory.django.DjangoModelFactory):
     """Factory for the ``Hgnc`` model."""
+
+    @classmethod
+    def _create(cls, model_class, *args, **kwargs):
+        """Create corresponding RefseqToHgnc object to the current Hgnc object."""
+        manager = cls._get_manager(model_class)
+        RefseqToHgncFactory(entrez_id=kwargs["entrez_id"])
+        return manager.create(*args, **kwargs)
 
     class Meta:
         model = Hgnc
@@ -14,8 +31,8 @@ class HgncFactory(factory.django.DjangoModelFactory):
     hgnc_id = factory.Sequence(lambda n: "HGNC:%d" % n)
     symbol = factory.Sequence(lambda n: "SYMBOL%d" % n)
     name = factory.Sequence(lambda n: "Gene name %d" % n)
-    entrez_id = factory.Sequence(lambda n: str(n))
     ensembl_gene_id = factory.Sequence(lambda n: "ENSG%d" % n)
+    entrez_id = factory.Sequence(lambda n: str(n))
 
 
 class HpoFactory(factory.django.DjangoModelFactory):
@@ -36,3 +53,14 @@ class Mim2geneMedgenFactory(factory.django.DjangoModelFactory):
 
     omim_id = factory.Sequence(lambda n: n)
     entrez_id = factory.Sequence(lambda n: str(n))
+
+
+class AcmgFactory(factory.django.DjangoModelFactory):
+    """Factory for the ``Acmg`` model."""
+
+    class Meta:
+        model = Acmg
+
+    entrez_id = factory.Sequence(lambda n: str(n))
+    ensembl_gene_id = factory.Sequence(lambda n: "ENSG%d" % n)
+    symbol = factory.Sequence(lambda n: "SYMBOL%d" % n)